Changed data type for price from money to double in DB and the Ticket entity. Got into trouble with the Client field in Ticket entity.

+++ sql/fill_objects.sql
@@ -16,7 +16,7 @@ create sequence ticket_id_seq;
 create table ticket (
   id integer not null default nextval('ticket_id_seq'),
   date date,
-  price money,
+  price double precision,
   client_id int,
   constraint pk_ticket primary key (id),
   constraint fk_ticket_client foreign key (client_id) references client (id)
